Move dedicated components to MainPage Header folder in UI
