--- ui/apps/platform/src/Components/ClusterStatusProblems/index.ts
@@ -1,1 +1,0 @@
-export { default } from './ClusterStatusProblems';
