+++ ui/apps/platform/src/Containers/MainPage/Header/GlobalSearchButton.tsx
